# Conflicts:
#	src/main/java/project/controllers/DefaultController.java
#	src/main/java/project/controllers/rest/ApiAuthController.java

+++ src/main/java/project/repositories/UserRepository.java
@@ -5,5 +5,6 @@ import org.springframework.stereotype.Repository;
 import project.models.User;
 
 @Repository
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, Integer> {
+    
 }
